Feature/search (#95)
* Search 추가

- 키워드별 검색
- 조건검색 추가

* Search 추가

- 리뷰 평점 추가
- total count 추가

+++ src/main/java/com/alconn/copang/item/dto/ItemDetailForm.java
@@ -1,11 +1,10 @@
 package com.alconn.copang.item.dto;
 
 
+import com.alconn.copang.shipment.dto.ShipmentInfoForm;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
-
-import com.alconn.copang.shipment.dto.ShipmentInfoForm;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -65,7 +64,7 @@ public class ItemDetailForm {
 
         private Long sellerId;
 
-        private Integer averageRating;
+        private Double averageRating;
 
         private ShipmentInfoForm shipmentInfoForm;
     }
